:recycle: Rename interfaces to types
